requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
array-record==0.4.1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
curl_cffi==0.6.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
Django==5.0.4
dm-tree==0.1.8
dukpy==0.4.0
etils==1.6.0
exceptiongroup==1.1.3
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.1
fsspec==2023.12.2
g4f==0.2.7.2
gast==0.4.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
image==1.5.33
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.6.1
jedi==0.19.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langchain-core==0.2.1
langchain-openai==0.1.7
langsmith==0.1.63
libclang==16.0.6
lxml==5.2.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.5
mutagen==1.47.0
mutf8==1.0.6
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
openai==1.30.3
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.3
patsy==0.5.6
Pillow==10.1.0
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
regex==2024.5.15
requests==2.32.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
stack-data==0.6.3
statsmodels==0.14.2
streamlink==6.8.1
tenacity==8.3.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.22.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trio==0.25.0
trio-websocket==0.11.1
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.1.0
wcwidth==0.2.9
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
you-get==0.4.1710
yt-dlp==2024.5.27
zipp==3.17.0
